-create separate classes for connection and data threads
