+++ .gitignore
@@ -13,3 +13,4 @@
 /captures
 .externalNativeBuild
 .cxx
+.idea/gradle.xml
